Export Fusco types as fusco:TYPE/ARITY
Exporting types by a header file is a poor idea as they use up the
global namespace. Take the type `result/0` for example; what if some
other application (that has to be used in one source file with fusco)
also wanted to use a type of that name but different
definition and exported it via a header file?
We would include both header file and voila -- name clash!